+++ ionoscloud/resource_datacenter.go
@@ -47,7 +47,7 @@ func resourceDatacenter() *schema.Resource {
 				Computed: true,
 			},
 			"features": {
-				Type:     schema.TypeList,
+				Type:     schema.TypeSet,
 				Computed: true,
 				Elem: &schema.Schema{
 					Type: schema.TypeString,